backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-CORS==4.0.1
APScheduler==3.10.4
requests==2.32.3
beautifulsoup4==4.12.3
selenium==4.25.0
python-dotenv==1.0.1
gunicorn==23.0.0
